refactor(hmi): split god-files, de-IIFE, addEventListener wiring; ADR: no state library
Phase 3 final slice + Phase 4 decision (PR 7 of epic #59).

app.ts (1271 lines) split along its seams:
- src/smr/state.ts: shared SMR state (graph, bounds, viewBox, ILS) +
  toSVG/attrNum helpers.
- src/smr/render.ts: airport SVG, ILS centerline, zoom-invariant
  rescaling, Aurora labels, updateSMRAircraft.
- src/smr/interaction.ts: pan/zoom + middle-drag label repositioning.
- src/polling.ts: the five refresh loops + app data (flightPlans,
  livePositions, currentICAO) — the app's de-facto store.
- src/rimcas.ts, src/runway-sequence.ts, src/lighting.ts.
- src/main.ts is now the composition root.

ptt.ts: the 47-complexity _onRecordingDone is a named-step pipeline
(_encodeRecording -> transcribe -> _pushControllerMessage ->
_dispatchAndRenderReply); chat bubble rendering extracted to
src/chat/ui.ts, shared with the WebSocket client.

De-IIFE: atis, debrief, setup, asr and ptt are plain modules with named
exports. Every inline on*= handler in both HTML pages is replaced by
addEventListener wiring in each module's init; the window bridges and
the transitional Window declarations are gone (only HMI_CONFIG stays).
The DOM-contract test now enforces the absence of inline handlers and
scans all of src/.

Phase 4: docs/adr-001-no-global-state-library.md records the evaluated
decision NOT to introduce Redux Toolkit (no React; polling.ts already
centralizes ownership) and the triggers under which RTK becomes
mandatory per the project rule.

Verified: 53 vitest + typecheck + lint + build green; 184 pytest green;
TestClient serves both pages with zero inline handlers.
