+++ services/controller_hmi_service/frontend/src/legacy/asr.ts
@@ -2,7 +2,7 @@
 // AI backend selection.
 
 import { getAsrSettings, setAsrSettings, type AsrSettings } from '../lib/storage';
-import { App } from './setup';
+import { showSession } from './setup';
 
 // ASR endpoint is injected at runtime via window.HMI_CONFIG (generated by
 // controller_hmi_service/main.py from the ASR_URL env var). Falls back to
@@ -19,382 +19,382 @@ function getAsrApi(): string {
 
 const ASR_API = getAsrApi();
 
-const Asr = (() => {
-  let _cfg: AsrSettings = {};
-  let _capturingPtt = false;
-  let _pttCaptureHandler: ((e: KeyboardEvent) => void) | null = null;
+let _cfg: AsrSettings = {};
+let _capturingPtt = false;
+let _pttCaptureHandler: ((e: KeyboardEvent) => void) | null = null;
 
-  // VU meter state
-  let _audioCtx: AudioContext | null = null;
-  let _analyser: AnalyserNode | null = null;
-  let _micStream: MediaStream | null = null;
-  let _vuRafId: number | null = null;
+// VU meter state
+let _audioCtx: AudioContext | null = null;
+let _analyser: AnalyserNode | null = null;
+let _micStream: MediaStream | null = null;
+let _vuRafId: number | null = null;
 
-  function _input(id: string): HTMLInputElement | null {
-    return document.getElementById(id) as HTMLInputElement | null;
+function _input(id: string): HTMLInputElement | null {
+  return document.getElementById(id) as HTMLInputElement | null;
+}
+
+function _select(id: string): HTMLSelectElement | null {
+  return document.getElementById(id) as HTMLSelectElement | null;
+}
+
+// ------------------------------------------------------------------
+// Screen init (called by setup.showAsr)
+// ------------------------------------------------------------------
+
+export async function initAsrScreen(): Promise<void> {
+  _clearError();
+  _cfg = getAsrSettings();
+  await _populateDevices();
+  _applyConfigToUI();
+  if (_cfg.backend === 'ollama') {
+    await _checkOllamaStatus();
+  }
+  _startVuMeter();
+}
+
+// Called when navigating away — stop mic
+export function teardownAsr(): void {
+  _stopVuMeter();
+  _stopPttCapture();
+}
+
+// ------------------------------------------------------------------
+// Config save
+// ------------------------------------------------------------------
+
+export async function saveAsrConfig(): Promise<void> {
+  _clearError();
+  teardownAsr();
+
+  const backend =
+    document.querySelector<HTMLInputElement>('input[name="asr-backend"]:checked')?.value ||
+    'ollama';
+  const payload: AsrSettings = {
+    input_device: _select('asr-input-device')?.value,
+    output_device: _select('asr-output-device')?.value,
+    ptt_key: document.getElementById('asr-ptt-btn')?.dataset.key || 'Space',
+    backend,
+    ollama_url: _input('asr-ollama-url')?.value.trim(),
+    ollama_model: _select('asr-ollama-model')?.value,
+    api_model: _select('asr-api-model')?.value || 'medium',
+  };
+
+  setAsrSettings(payload);
+
+  _cfg = payload;
+  showSession();
+}
+
+// ------------------------------------------------------------------
+// Populate UI from _cfg
+// ------------------------------------------------------------------
+
+function _applyConfigToUI(): void {
+  const backend = _cfg.backend || 'ollama';
+
+  const radio = document.querySelector<HTMLInputElement>(
+    `input[name="asr-backend"][value="${backend}"]`,
+  );
+  if (radio) radio.checked = true;
+  switchBackend(backend);
+
+  const pttKey = _cfg.ptt_key || 'Space';
+  const pttBtn = document.getElementById('asr-ptt-btn');
+  if (pttBtn) {
+    pttBtn.textContent = pttKey;
+    pttBtn.dataset.key = pttKey;
   }
 
-  function _select(id: string): HTMLSelectElement | null {
-    return document.getElementById(id) as HTMLSelectElement | null;
+  if (_cfg.ollama_url) {
+    const el = _input('asr-ollama-url');
+    if (el) el.value = _cfg.ollama_url;
+  }
+  if (_cfg.ollama_model) {
+    const el = _select('asr-ollama-model');
+    if (el) el.value = _cfg.ollama_model;
+  }
+  if (_cfg.api_key) {
+    const el = _input('asr-api-key');
+    if (el) el.value = _cfg.api_key;
+  }
+  if (_cfg.api_base_url) {
+    const el = _input('asr-api-base');
+    if (el) el.value = _cfg.api_base_url;
+  }
+  if (_cfg.api_model) {
+    const el = _select('asr-api-model');
+    if (el) el.value = _cfg.api_model;
   }
 
-  // ------------------------------------------------------------------
-  // Screen init (called by App.showAsr)
-  // ------------------------------------------------------------------
+  if (_cfg.input_device) {
+    const el = _select('asr-input-device');
+    if (el && [...el.options].some((o) => o.value === _cfg.input_device))
+      el.value = _cfg.input_device;
+  }
+  if (_cfg.output_device) {
+    const el = _select('asr-output-device');
+    if (el && [...el.options].some((o) => o.value === _cfg.output_device))
+      el.value = _cfg.output_device;
+  }
+}
 
-  async function initScreen(): Promise<void> {
-    _clearError();
-    _loadConfig();
-    await _populateDevices();
-    _applyConfigToUI();
-    if (_cfg.backend === 'ollama') {
-      await _checkOllamaStatus();
-    }
-    _startVuMeter();
+// ------------------------------------------------------------------
+// Audio devices
+// ------------------------------------------------------------------
+
+async function _populateDevices(): Promise<void> {
+  try {
+    const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
+    stream.getTracks().forEach((t) => t.stop());
+  } catch {
+    /* labels may be empty without permission */
   }
 
-  // Called when App navigates away — stop mic
-  function teardown(): void {
-    _stopVuMeter();
-    _stopPttCapture();
+  try {
+    const devices = await navigator.mediaDevices.enumerateDevices();
+    const inputSel = _select('asr-input-device');
+    const outputSel = _select('asr-output-device');
+    if (!inputSel || !outputSel) return;
+
+    inputSel.innerHTML = '<option value="default">Default</option>';
+    outputSel.innerHTML = '<option value="default">Default</option>';
+
+    devices.forEach((d) => {
+      const opt = document.createElement('option');
+      opt.value = d.deviceId;
+      opt.textContent = d.label || d.deviceId.slice(0, 24);
+
+      if (d.kind === 'audioinput') inputSel.appendChild(opt.cloneNode(true));
+      if (d.kind === 'audiooutput') outputSel.appendChild(opt);
+    });
+  } catch {
+    /* no device access */
   }
+}
 
-  // ------------------------------------------------------------------
-  // Config load / save
-  // ------------------------------------------------------------------
+// Called when user switches input device — restart VU meter on new device
+export async function onInputDeviceChange(): Promise<void> {
+  _stopVuMeter();
+  await _startVuMeter();
+}
 
-  function _loadConfig(): void {
-    _cfg = getAsrSettings();
-  }
+// ------------------------------------------------------------------
+// VU Meter (Web Audio API)
+// ------------------------------------------------------------------
 
-  async function saveConfig(): Promise<void> {
-    _clearError();
-    teardown();
+async function _startVuMeter(): Promise<void> {
+  _stopVuMeter();
 
-    const backend =
-      document.querySelector<HTMLInputElement>('input[name="asr-backend"]:checked')?.value ||
-      'ollama';
-    const payload: AsrSettings = {
-      input_device: _select('asr-input-device')?.value,
-      output_device: _select('asr-output-device')?.value,
-      ptt_key: document.getElementById('asr-ptt-btn')?.dataset.key || 'Space',
-      backend,
-      ollama_url: _input('asr-ollama-url')?.value.trim(),
-      ollama_model: _select('asr-ollama-model')?.value,
-      api_model: _select('asr-api-model')?.value || 'medium',
-    };
+  const deviceId = _select('asr-input-device')?.value;
+  const constraints: MediaStreamConstraints = {
+    audio: deviceId && deviceId !== 'default' ? { deviceId: { ideal: deviceId } } : true,
+  };
 
-    setAsrSettings(payload);
+  try {
+    _micStream = await navigator.mediaDevices.getUserMedia(constraints);
+    _audioCtx = new AudioContext();
+    if (_audioCtx.state === 'suspended') await _audioCtx.resume();
+    _analyser = _audioCtx.createAnalyser();
+    _analyser.fftSize = 256;
+    _analyser.smoothingTimeConstant = 0.6;
 
-    _cfg = payload;
-    App.showSession();
-  }
-
-  // ------------------------------------------------------------------
-  // Populate UI from _cfg
-  // ------------------------------------------------------------------
-
-  function _applyConfigToUI(): void {
-    const backend = _cfg.backend || 'ollama';
-
-    const radio = document.querySelector<HTMLInputElement>(
-      `input[name="asr-backend"][value="${backend}"]`,
-    );
-    if (radio) radio.checked = true;
-    switchBackend(backend);
-
-    const pttKey = _cfg.ptt_key || 'Space';
-    const pttBtn = document.getElementById('asr-ptt-btn');
-    if (pttBtn) {
-      pttBtn.textContent = pttKey;
-      pttBtn.dataset.key = pttKey;
-    }
-
-    if (_cfg.ollama_url) {
-      const el = _input('asr-ollama-url');
-      if (el) el.value = _cfg.ollama_url;
-    }
-    if (_cfg.ollama_model) {
-      const el = _select('asr-ollama-model');
-      if (el) el.value = _cfg.ollama_model;
-    }
-    if (_cfg.api_key) {
-      const el = _input('asr-api-key');
-      if (el) el.value = _cfg.api_key;
-    }
-    if (_cfg.api_base_url) {
-      const el = _input('asr-api-base');
-      if (el) el.value = _cfg.api_base_url;
-    }
-    if (_cfg.api_model) {
-      const el = _select('asr-api-model');
-      if (el) el.value = _cfg.api_model;
-    }
-
-    if (_cfg.input_device) {
-      const el = _select('asr-input-device');
-      if (el && [...el.options].some((o) => o.value === _cfg.input_device))
-        el.value = _cfg.input_device;
-    }
-    if (_cfg.output_device) {
-      const el = _select('asr-output-device');
-      if (el && [...el.options].some((o) => o.value === _cfg.output_device))
-        el.value = _cfg.output_device;
-    }
-  }
-
-  // ------------------------------------------------------------------
-  // Audio devices
-  // ------------------------------------------------------------------
-
-  async function _populateDevices(): Promise<void> {
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-      stream.getTracks().forEach((t) => t.stop());
-    } catch {
-      /* labels may be empty without permission */
-    }
-
-    try {
-      const devices = await navigator.mediaDevices.enumerateDevices();
-      const inputSel = _select('asr-input-device');
-      const outputSel = _select('asr-output-device');
-      if (!inputSel || !outputSel) return;
-
-      inputSel.innerHTML = '<option value="default">Default</option>';
-      outputSel.innerHTML = '<option value="default">Default</option>';
-
-      devices.forEach((d) => {
-        const opt = document.createElement('option');
-        opt.value = d.deviceId;
-        opt.textContent = d.label || d.deviceId.slice(0, 24);
-
-        if (d.kind === 'audioinput') inputSel.appendChild(opt.cloneNode(true));
-        if (d.kind === 'audiooutput') outputSel.appendChild(opt);
-      });
-    } catch {
-      /* no device access */
-    }
-  }
-
-  // Called when user switches input device — restart VU meter on new device
-  async function onInputDeviceChange(): Promise<void> {
-    _stopVuMeter();
-    await _startVuMeter();
-  }
-
-  // ------------------------------------------------------------------
-  // VU Meter (Web Audio API)
-  // ------------------------------------------------------------------
-
-  async function _startVuMeter(): Promise<void> {
-    _stopVuMeter();
-
-    const deviceId = _select('asr-input-device')?.value;
-    const constraints: MediaStreamConstraints = {
-      audio: deviceId && deviceId !== 'default' ? { deviceId: { ideal: deviceId } } : true,
-    };
-
-    try {
-      _micStream = await navigator.mediaDevices.getUserMedia(constraints);
-      _audioCtx = new AudioContext();
-      if (_audioCtx.state === 'suspended') await _audioCtx.resume();
-      _analyser = _audioCtx.createAnalyser();
-      _analyser.fftSize = 256;
-      _analyser.smoothingTimeConstant = 0.6;
-
-      const source = _audioCtx.createMediaStreamSource(_micStream);
-      source.connect(_analyser);
-
-      _vuRafId = requestAnimationFrame(_vuTick);
-      _setVuHint('active');
-    } catch {
-      _setVuHint('no mic');
-    }
-  }
-
-  function _stopVuMeter(): void {
-    if (_vuRafId) {
-      cancelAnimationFrame(_vuRafId);
-      _vuRafId = null;
-    }
-    if (_analyser) {
-      _analyser.disconnect();
-      _analyser = null;
-    }
-    if (_audioCtx) {
-      _audioCtx.close();
-      _audioCtx = null;
-    }
-    if (_micStream) {
-      _micStream.getTracks().forEach((t) => t.stop());
-      _micStream = null;
-    }
-    const bar = document.getElementById('asr-vu-bar');
-    if (bar) bar.style.width = '0%';
-    _setVuHint('—');
-  }
-
-  function _vuTick(): void {
-    if (!_analyser) return;
-    const data = new Uint8Array(_analyser.frequencyBinCount);
-    _analyser.getByteFrequencyData(data);
-
-    // RMS of frequency bins → 0..1
-    const rms = Math.sqrt(data.reduce((s, v) => s + v * v, 0) / data.length) / 128;
-    const pct = Math.min(100, Math.round(rms * 180));
-
-    const bar = document.getElementById('asr-vu-bar');
-    if (bar) bar.style.width = pct + '%';
-    _setVuHint(pct + '%');
+    const source = _audioCtx.createMediaStreamSource(_micStream);
+    source.connect(_analyser);
 
     _vuRafId = requestAnimationFrame(_vuTick);
+    _setVuHint('active');
+  } catch {
+    _setVuHint('no mic');
   }
+}
 
-  function _setVuHint(text: string): void {
-    const el = document.getElementById('asr-vu-hint');
-    if (el) el.textContent = text;
+function _stopVuMeter(): void {
+  if (_vuRafId) {
+    cancelAnimationFrame(_vuRafId);
+    _vuRafId = null;
   }
+  if (_analyser) {
+    _analyser.disconnect();
+    _analyser = null;
+  }
+  if (_audioCtx) {
+    _audioCtx.close();
+    _audioCtx = null;
+  }
+  if (_micStream) {
+    _micStream.getTracks().forEach((t) => t.stop());
+    _micStream = null;
+  }
+  const bar = document.getElementById('asr-vu-bar');
+  if (bar) bar.style.width = '0%';
+  _setVuHint('—');
+}
 
-  // ------------------------------------------------------------------
-  // Ollama models + status LED
-  // ------------------------------------------------------------------
+function _vuTick(): void {
+  if (!_analyser) return;
+  const data = new Uint8Array(_analyser.frequencyBinCount);
+  _analyser.getByteFrequencyData(data);
 
-  async function _checkOllamaStatus(): Promise<string[]> {
-    _setLed('checking');
-    try {
-      const res = await fetch(`${ASR_API}/ollama/models`);
-      if (res.ok) {
-        const data = (await res.json()) as { models?: string[] };
-        _setLed('ok');
-        return data.models || [];
-      }
-      _setLed('err');
-      return [];
-    } catch {
-      _setLed('err');
-      return [];
+  // RMS of frequency bins → 0..1
+  const rms = Math.sqrt(data.reduce((s, v) => s + v * v, 0) / data.length) / 128;
+  const pct = Math.min(100, Math.round(rms * 180));
+
+  const bar = document.getElementById('asr-vu-bar');
+  if (bar) bar.style.width = pct + '%';
+  _setVuHint(pct + '%');
+
+  _vuRafId = requestAnimationFrame(_vuTick);
+}
+
+function _setVuHint(text: string): void {
+  const el = document.getElementById('asr-vu-hint');
+  if (el) el.textContent = text;
+}
+
+// ------------------------------------------------------------------
+// Ollama models + status LED
+// ------------------------------------------------------------------
+
+async function _checkOllamaStatus(): Promise<string[]> {
+  _setLed('checking');
+  try {
+    const res = await fetch(`${ASR_API}/ollama/models`);
+    if (res.ok) {
+      const data = (await res.json()) as { models?: string[] };
+      _setLed('ok');
+      return data.models || [];
     }
+    _setLed('err');
+    return [];
+  } catch {
+    _setLed('err');
+    return [];
+  }
+}
+
+async function _populateOllamaModels(): Promise<void> {
+  const sel = _select('asr-ollama-model');
+  if (!sel) return;
+  const current = sel.value || _cfg.ollama_model || 'whisper';
+  const models = await _checkOllamaStatus();
+  if (models.length === 0) return;
+  sel.innerHTML = '';
+  models.forEach((m) => {
+    const opt = document.createElement('option');
+    opt.value = m;
+    opt.textContent = m;
+    if (m === current) opt.selected = true;
+    sel.appendChild(opt);
+  });
+}
+
+export async function reloadOllamaModels(): Promise<void> {
+  await _populateOllamaModels();
+}
+
+function _setLed(state: string): void {
+  const led = document.getElementById('asr-ollama-led');
+  if (!led) return;
+  led.className = 'status-led' + (state ? ' ' + state : '');
+}
+
+// ------------------------------------------------------------------
+// Backend toggle
+// ------------------------------------------------------------------
+
+export function switchBackend(val: string): void {
+  document.getElementById('asr-sec-ollama')?.classList.toggle('hidden', val !== 'ollama');
+  document.getElementById('asr-sec-api')?.classList.toggle('hidden', val !== 'api');
+  if (val === 'ollama') _populateOllamaModels();
+  else _setLed('');
+}
+
+// ------------------------------------------------------------------
+// PTT key capture
+// ------------------------------------------------------------------
+
+export function captureAsrPtt(): void {
+  if (_capturingPtt) return;
+  _capturingPtt = true;
+
+  const btn = document.getElementById('asr-ptt-btn');
+  if (btn) {
+    btn.textContent = 'Press a key...';
+    btn.classList.add('listening');
   }
 
-  async function _populateOllamaModels(): Promise<void> {
-    const sel = _select('asr-ollama-model');
-    if (!sel) return;
-    const current = sel.value || _cfg.ollama_model || 'whisper';
-    const models = await _checkOllamaStatus();
-    if (models.length === 0) return;
-    sel.innerHTML = '';
-    models.forEach((m) => {
-      const opt = document.createElement('option');
-      opt.value = m;
-      opt.textContent = m;
-      if (m === current) opt.selected = true;
-      sel.appendChild(opt);
-    });
-  }
-
-  async function reloadOllamaModels(): Promise<void> {
-    await _populateOllamaModels();
-  }
-
-  function _setLed(state: string): void {
-    const led = document.getElementById('asr-ollama-led');
-    if (!led) return;
-    led.className = 'status-led' + (state ? ' ' + state : '');
-  }
-
-  // ------------------------------------------------------------------
-  // Backend toggle
-  // ------------------------------------------------------------------
-
-  function switchBackend(val: string): void {
-    document.getElementById('asr-sec-ollama')?.classList.toggle('hidden', val !== 'ollama');
-    document.getElementById('asr-sec-api')?.classList.toggle('hidden', val !== 'api');
-    if (val === 'ollama') _populateOllamaModels();
-    else _setLed('');
-  }
-
-  // ------------------------------------------------------------------
-  // PTT key capture
-  // ------------------------------------------------------------------
-
-  function capturePtt(): void {
-    if (_capturingPtt) return;
-    _capturingPtt = true;
-
-    const btn = document.getElementById('asr-ptt-btn');
+  _pttCaptureHandler = (e: KeyboardEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    const key = e.code || e.key;
     if (btn) {
-      btn.textContent = 'Press a key...';
-      btn.classList.add('listening');
+      btn.textContent = _keyLabel(key);
+      btn.dataset.key = key;
     }
-
-    _pttCaptureHandler = (e: KeyboardEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      const key = e.code || e.key;
-      if (btn) {
-        btn.textContent = _keyLabel(key);
-        btn.dataset.key = key;
-      }
-      _stopPttCapture();
-    };
-
-    window.addEventListener('keydown', _pttCaptureHandler, { capture: true, once: true });
-  }
-
-  function _stopPttCapture(): void {
-    if (_pttCaptureHandler) {
-      window.removeEventListener('keydown', _pttCaptureHandler, { capture: true });
-      _pttCaptureHandler = null;
-    }
-    _capturingPtt = false;
-    const btn = document.getElementById('asr-ptt-btn');
-    if (btn) btn.classList.remove('listening');
-  }
-
-  function _keyLabel(code: string): string {
-    const map: Record<string, string> = {
-      Space: 'Space',
-      Enter: 'Enter',
-      Escape: 'Esc',
-      ShiftLeft: 'L-Shift',
-      ShiftRight: 'R-Shift',
-      ControlLeft: 'L-Ctrl',
-      ControlRight: 'R-Ctrl',
-      AltLeft: 'L-Alt',
-      AltRight: 'R-Alt',
-      CapsLock: 'Caps',
-    };
-    const mapped = map[code];
-    if (mapped) return mapped;
-    if (code.startsWith('Key')) return code.slice(3);
-    if (code.startsWith('Digit')) return code.slice(5);
-    if (code.startsWith('Numpad')) return 'NP' + code.slice(6);
-    return code;
-  }
-
-  // ------------------------------------------------------------------
-  // Error helpers
-  // ------------------------------------------------------------------
-
-  function _clearError(): void {
-    const el = document.getElementById('asr-error');
-    if (!el) return;
-    el.textContent = '';
-    el.classList.add('hidden');
-  }
-
-  return {
-    initScreen,
-    teardown,
-    saveConfig,
-    switchBackend,
-    capturePtt,
-    reloadOllamaModels,
-    onInputDeviceChange,
+    _stopPttCapture();
   };
-})();
 
-// Bridge for the inline on*= handlers in setup.html (removed in Phase 3).
-window.Asr = Asr;
-export { Asr };
+  window.addEventListener('keydown', _pttCaptureHandler, { capture: true, once: true });
+}
+
+function _stopPttCapture(): void {
+  if (_pttCaptureHandler) {
+    window.removeEventListener('keydown', _pttCaptureHandler, { capture: true });
+    _pttCaptureHandler = null;
+  }
+  _capturingPtt = false;
+  const btn = document.getElementById('asr-ptt-btn');
+  if (btn) btn.classList.remove('listening');
+}
+
+function _keyLabel(code: string): string {
+  const map: Record<string, string> = {
+    Space: 'Space',
+    Enter: 'Enter',
+    Escape: 'Esc',
+    ShiftLeft: 'L-Shift',
+    ShiftRight: 'R-Shift',
+    ControlLeft: 'L-Ctrl',
+    ControlRight: 'R-Ctrl',
+    AltLeft: 'L-Alt',
+    AltRight: 'R-Alt',
+    CapsLock: 'Caps',
+  };
+  const mapped = map[code];
+  if (mapped) return mapped;
+  if (code.startsWith('Key')) return code.slice(3);
+  if (code.startsWith('Digit')) return code.slice(5);
+  if (code.startsWith('Numpad')) return 'NP' + code.slice(6);
+  return code;
+}
+
+// ------------------------------------------------------------------
+// Error helpers
+// ------------------------------------------------------------------
+
+function _clearError(): void {
+  const el = document.getElementById('asr-error');
+  if (!el) return;
+  el.textContent = '';
+  el.classList.add('hidden');
+}
+
+// ------------------------------------------------------------------
+// Wiring (replaces the former inline on*= handlers)
+// ------------------------------------------------------------------
+
+export function initAsrWiring(): void {
+  document
+    .getElementById('asr-input-device')
+    ?.addEventListener('change', () => void onInputDeviceChange());
+  document.getElementById('asr-ptt-btn')?.addEventListener('click', captureAsrPtt);
+  document
+    .getElementById('asr-refresh-btn')
+    ?.addEventListener('click', () => void reloadOllamaModels());
+  document.querySelectorAll<HTMLInputElement>('input[name="asr-backend"]').forEach((radio) => {
+    radio.addEventListener('change', () => switchBackend(radio.value));
+  });
+  document.getElementById('asr-save-btn')?.addEventListener('click', () => void saveAsrConfig());
+}
+
+document.addEventListener('DOMContentLoaded', initAsrWiring);
