+++ services/controller_hmi_service/frontend/src/main.ts
@@ -1,16 +1,27 @@
-// Entry point for index.html (TWR workstation).
-// Ordered side-effect imports mirror the original <script> tag order.
+// Composition root for index.html (TWR workstation).
+// Side-effect imports wire each panel's own listeners; the explicit init
+// below starts the clock, widgets, SMR map and polling loops.
 
 import './styles/efs.css';
 import './styles/debrief.css';
 
-import './legacy/weather';
-import './legacy/wind';
-import './legacy/efs';
 import './legacy/resize';
 import './legacy/atis';
 import './legacy/debrief';
 import './legacy/ptt';
 import './legacy/chat';
-import './legacy/app';
 import './legacy/header-widgets';
+
+import { initWindInstruments } from './legacy/wind';
+import { startUTCClock } from './legacy/weather';
+import { initLightingControls } from './lighting';
+import { startPolling } from './polling';
+import { initSMRMap } from './smr/render';
+
+document.addEventListener('DOMContentLoaded', () => {
+  startUTCClock();
+  initWindInstruments();
+  initLightingControls();
+  initSMRMap();
+  startPolling();
+});
